mnesia_to_khepri: Speed `is_migration_finished()` up
[Why]
The current approach queries Khepri to determine if a migration is in
progress or even finished. This is quite expensive for what is expected
a quick check to decide between the Mnesia and Khepri functions.

[How]
We use a Khepri projection and an ETS table to cache the migration
statuses. This way, the check simply becomes an ETS lookup.