mnesia_to_khepri: Support stopped Khepri store in `is_migration_finished/2`
[Why]
`is_migration_finished/{1,2}` and `handle_fallback/{3,4}` could be used
before Khepri is even configured. Therefore, it should support the fact
that the Khepri store might be stopped.

Without this, we get an exception because we try to set a projection up
with an unavailable store.

[How]
We look up the name of the given store in the list of running stores.