mnesia_to_khepri: Wait for projection to be deleted in `rollback/2`
[Why]
The projection of the migration marker is updated asynchronously.
Therefore it's possible that `is_migration_finished/2` still returns
true after a call to `rollback/2` for a short period of time.

[How]
After we clear the migration marker from Khepri, we wait for the
projection to be updated.

This makes sure that when `rollback/2` returns,
`is_migration_finished/2` returns false.

This fixes a race condition in the testsuite (which led to frequent
transient failures in CI). It's also more intuitive for users.