mnesia_to_khepri: Protect against infinite loop in `handle_fallback/4`
[Why]
If the Mnesia function fails because a Mnesia table does not exist for
real and this error is unrelated to a migration to Khepri,
`handle_fallback/4` will enter an infinite loop with a quick repetition.

With debug messages enabled, this causes many many stacktraces to be
logged.

[How]
We just count the number of times the Mnesia function fails with a
`no_exists` reason. After 100 attempts, we stop retrying.