m2k_table_copy: Handle existing projection
[Why]
Thanks to concurrency, it's possible that the projection is created
between the attempt to use it and the call to `setup_projection/2`. If
this happens, the setup fails with an error telling the projection
already exists.

This is not a fatal error, we should simply retry to read the
projection's table.

[How]
We handle both the old and the new possible "projection already exists"
error terms and retry the function if this occurs.