m2k_table_copy: Reword projection setup failure warning
... and reduce its severity from "warning" to "info".

[Why]
The warning, being logged with every call to `handle_fallback()` where
we couldn't set the migration up, was too scary. Even more when it
happened when Khepri was not being used yet.

[How]
The severity is decreased to "info" because in the end,
`handle_fallback()` will work equally well, just slower.

The message is rephrased to indicate that it's part of the database
selection, not the table copy. It also states that the issue is
harmless.