m2k_table_copy: Log migration projection setup failure once
[Why]
If we fail to setup the projection, we used to log the error for every
calls. Because `is_migration_finished/2` and `handle_fallback/4` are
meant to be used a lot, we could spam log messages a lot.

This is not a critical error as the function can work with a regular
Khepri query. There is no need to log so much.

[How]
Now, a given error is logged once only. If the error changes, we issue
another log message once.