m2k_cluster_sync: Tolerate that a node is down during sync
[Why]
In commit 52d3e5660f752d5a533af91906db7ffb0c22d73b, we handled the
situation where a node lost its data and we tried to sync the cluster.

This patch was not enough if we lose sevenal nodes because some of them
might be still down, or the Khepri store might not run yet. In this
case, the sync fails because it can't remove those nodes from the
cluster.

[How]
This patch does two things in `remove_nodes_from_khepri_cluster/2`:
* It switches to use `erpc` and catches the fact that the remote node
  could be unreachable.
* It handles the `not_a_khepri_store` error.

In both cases, it ignores and skips that node which should have beer
removed and just log a message.