m2k_table_copy: Improve logging during clean up
[Why]
The clean up can run concurrently on several nodes. Therefore a Mnesia
table may already be marked as read-write or deleted by another node. So
far, it cause the function to log an error. This is scary for no good
reason.

[How]
If the error is about a table that is already marked as read-write or
already deleted, we log that it is deleted and move on.