m2k_table_copy: Handle the "delete" backup items
[Why]
The Mnesia Backup and Restore callback API states that the callback
module must provide the following function:

    write(Priv, BackupItems)

`BackupItems` can be among other values:
* `{Record}` to pass a record that is present in the table
* `{Table, Key}` to pass a record that should be deleted

Unlike what the docs state, the passed record present in the table is
passed as `Record`; in other words, it is not wrapped in a tuple. This
is not really a problem because a size-2 record is not allowed to be
used to define a Mnesia table schema.

Our testsuite never met a case where the `{Table, Key}` to delete is
passed to the callback module. However it happens in RabbitMQ when a lot
of queues are declared in parallel of the migration.

Before this patch, khepri_mnesia_migration would consider this term an
invalid record and would ignore it. That is incorrect and the key should
be deleted instead.

[How]
Now, `m2k_table_copy` calls its own callback module's
`delete_from_khepri()` function if the backup item is `{Table, Key}`,
and calls `copy_to_khepri()` otherwise.