m2k_cluster_sync: Strengthen the conditions required to repair a cluster
[Why]
So far, when an Erlang node was not clustered according to Mnesia, we
still considered other connected Erlang nodes to determine if the local
node should be clustered (with them). This allowed to repair a cluster
where a node lost all its data.

There was a single condition used to filter connected nodes: they must
have been running the same Khepri store (based on its ID).

Unfortunately, this was too vague and led to the "repair" of clusters
that did not need any repair. For instance, if a node explicitly left a
cluster (or was kicked out of it) and was reset to start from scratch,
them khepri_mnesia_migration would re-add the node back to the cluster
because the condition was met.

[How]
The code enforces a second condition now: the connected nodes must have
the local node in their cluster members. This way, we are sure that
there is a disconnect between the state of the local node and what other
nodes think the cluster should look like.