cluster sync: Use `khepri_cluster:members/2` from Khepri 0.17
Khepri 0.17 replaces `khepri_cluster:locally_known_members/1` with a
new version of `khepri_cluster:nodes` which accepts a store and query
options. This change updates the RPC call to `locally_known_members/1`
to first try the new `members/2` function with the equivalent option
`#{favor => low_latency}` and falls back to `locally_known_members/1`
when this is undefined on the remote Erlang node.