m2k_cluster_sync: Add an application env. var. to disable the cluster repair mechanism
[Why]
The caller may want to forcibly reset a Khepri node without a clean
unclustering, leaving other nodes connected and thinking they are still
connected to the reset node.

This is the case of RabbitMQ with it "force reset" command.

[How]
This application environment variable allows it to prever the repair
mechanism to revert that reset and put the reset node back into the
cluster.