mnesia_to_khepri_cluster_sync_SUITE: Re-enable cluster repair tests
[Why]
They were disabled because of a transient failure documented in a
comment in the code: the lost node could receive an `append_entry` RPC
and it would exit because it didn't expect one.

[How]
We now wait for a leader election in the remaining cluster before
proceeding with the test. It looks to eliminate the chance of getting
this `append_entry`.