m2k_cluster_sync: Sort clusters by Khepri machine version
[Why]
If one of the node/cluster uses a lower Khepri machine version as other
nodes/clusters, it won't be able to join a node/cluster running a newer
version.

[How]
We now sort a node/cluster using the lower Khepri machine version first.
This way, it will be used as the seed node/cluster for the sync.

The members count remains a more important criteria, so the largest
cluster won't be dismantled to join a single node running an old Khepri
machine version.

However, other criterias are less important.